+++ src/core/parsers/bcovlive.ts
@@ -1,5 +1,0 @@
-import { ParserOptions, ParserResult } from "./types";
-
-export function parseBcovLive(_options: ParserOptions): ParserResult {
-    return { dropChunksOnMaxRetries: true };
-}